Unify local decoration tests
