+++ tests/Agent/IntegrationTests/IntegrationTests/Logging/Common.cs
@@ -1,0 +1,12 @@
+﻿// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+namespace NewRelic.Agent.IntegrationTests.Logging
+{
+    public enum LoggingFramework
+    {
+        Log4net,
+        Serilog,
+        MicrosoftLogging
+    }
+}
